+++ src/aks-preview/azext_aks_preview/commands.py
@@ -13,6 +13,7 @@ from azext_aks_preview._client_factory import (
     cf_nodepool_snapshots,
     cf_trustedaccess_role,
     cf_trustedaccess_role_binding,
+    cf_machines
 )
 from azext_aks_preview._format import (
     aks_addon_list_available_table_format,
@@ -20,6 +21,8 @@ from azext_aks_preview._format import (
     aks_addon_show_table_format,
     aks_agentpool_list_table_format,
     aks_agentpool_show_table_format,
+    aks_machine_list_table_format,
+    aks_machine_show_table_format,
     aks_list_nodepool_snapshot_table_format,
     aks_list_snapshot_table_format,
     aks_list_table_format,
@@ -85,6 +88,12 @@ def load_command_table(self, _):
     agent_pools_sdk = CliCommandType(
         operations_tmpl='azext_aks_preview.vendored_sdks.azure_mgmt_preview_aks.'
                         'operations._agent_pools_operations#AgentPoolsOperations.{}',
+        client_factory=cf_managed_clusters
+    )
+
+    machines_sdk = CliCommandType(
+        operations_tmpl='azext_aks_preview.vendored_sdks.azure_mgmt_preview_aks.'
+                        'operations._machine_operations#MachinesOperations.{}',
         client_factory=cf_managed_clusters
     )
 
@@ -184,6 +193,12 @@ def load_command_table(self, _):
         g.custom_command('stop', 'aks_agentpool_stop', supports_no_wait=True)
         g.custom_command('start', 'aks_agentpool_start', supports_no_wait=True)
         g.custom_command('operation-abort', 'aks_agentpool_operation_abort', supports_no_wait=True)
+
+    with self.command_group('aks machine', machines_sdk, client_factory=cf_machines) as g:
+        g.custom_command('list', 'aks_machine_list',
+                         table_transformer=aks_machine_list_table_format)
+        g.custom_show_command('show', 'aks_machine_show',
+                              table_transformer=aks_machine_show_table_format)
 
     # AKS draft commands
     with self.command_group('aks draft', managed_clusters_sdk, client_factory=cf_managed_clusters) as g: