+++ src/vmware/azext_vmware/tests/latest/test_vmware_scenario.py
@@ -22,7 +22,9 @@ class VmwareScenarioTest(ScenarioTest):
             'loc': 'centralus',
             'privatecloud': 'cloud1',
             'cluster': 'pycluster1',
-            'hosts': 'fakehost22.nyc1.kubernetes.center fakehost23.nyc1.kubernetes.center fakehost24.nyc1.kubernetes.center'
+            'hosts': 'fakehost22.nyc1.kubernetes.center fakehost23.nyc1.kubernetes.center fakehost24.nyc1.kubernetes.center',
+            'key_vault_key': 'vmwarekey',
+            'vault_url': 'https://keyvault1-kmip-kvault.vault.azure.net/'
         })
 
         # check quote availability
@@ -42,17 +44,15 @@ class VmwareScenarioTest(ScenarioTest):
         self.assertEqual(count, 1, 'private cloud count expected to be 1')
 
         # count at the subscription level
-        # test passes, but commented out for privacy
-        # count = len(self.cmd('vmware private-cloud list').get_output_in_json())
-        # self.assertGreaterEqual(count, 1, 'subscription private cloud count expected to be more than 1')
+        count = len(self.cmd('vmware private-cloud list').get_output_in_json())
+        self.assertGreaterEqual(count, 1, 'subscription private cloud count expected to be more than 1')
 
         # get admin credentials
-        # not currently supported in test environment
-        # self.cmd('vmware private-cloud listadmincredentials -g {rg} -c {privatecloud}')
+        self.cmd('vmware private-cloud listadmincredentials -g {rg} -c {privatecloud}')
 
         # rotate passwords
-        # self.cmd('vmware private-cloud rotate-vcenter-password -g {rg} -c {privatecloud}')
-        # self.cmd('vmware private-cloud rotate-nsxt-password -g {rg} -c {privatecloud}')
+        self.cmd('vmware private-cloud rotate-vcenter-password -g {rg} -c {privatecloud} --yes')
+        # self.cmd('vmware private-cloud rotate-nsxt-password')
 
         # update private cloud to changed default cluster size
         self.cmd('vmware private-cloud update -g {rg} -n {privatecloud} --cluster-size 4')
@@ -68,7 +68,7 @@ class VmwareScenarioTest(ScenarioTest):
 
         # add identity source
         self.cmd(
-            'vmware private-cloud add-identity-source -g {rg} -c {privatecloud} -n groupName --alias groupAlias --domain domain --base-user-dn "ou=baseUser" --base-group-dn "ou=baseGroup" --primary-server ldaps://1.1.1.1:636 --username someone --password something')
+            'vmware private-cloud add-identity-source -g {rg} -c {privatecloud} -n group1 --alias groupAlias --domain domain1 --base-user-dn "ou=baseUser" --base-group-dn "ou=baseGroup" --primary-server ldaps://1.1.1.1:636 --username someone --password something')
 
         # delete identity source
         self.cmd(
@@ -88,69 +88,17 @@ class VmwareScenarioTest(ScenarioTest):
         count = len(self.cmd('vmware cluster list -g {rg} -c {privatecloud}').get_output_in_json())
         self.assertEqual(count, 1, 'cluster count expected to be 1')
 
-        # # cluster update
-        # self.cmd('vmware cluster update -g {rg} -c {privatecloud} -n {cluster} --size 4')
+        # cluster update
+        self.cmd('vmware cluster update -g {rg} -c {privatecloud} -n {cluster} --size 4')
 
-        # cluster delete
-        self.cmd('vmware cluster delete -g {rg} -c {privatecloud} -n {cluster} --yes')
+        # create extended network blocks
+        self.cmd('vmware private-cloud update -g {rg} -n {privatecloud} --ext-nw-blocks 10.0.8.0/23 10.1.8.0/23 10.2.8.0/23')
 
-        # cluster list zone
-        self.cmd('vmware cluster list-zones -g {rg} -c {privatecloud} -n {cluster}')
-
-        # delete the private cloud
-        self.cmd('vmware private-cloud delete -g {rg} -n {privatecloud} --yes')
-
-    @record_only()
-    def test_vmware_private_cloud(self):
-        self.kwargs.update({
-            'rg': "cli_test_vmwarepc",
-            "kv": "cli-test-vmwarepc0821",
-            'privatecloud': 'cloud1',
-            'sku': 'av36',
-            # 'cluster': 'pycluster1',
-            'key_vault_key': "vmwarekey",
-            'loc': "brazilsouth"
-        })
-
-        self.cmd("group create --name {rg} --location {loc}")
-
-        # check quote availability
-        self.cmd('vmware location check-quota-availability --location {loc}')
-
-        # check trial availability
-        self.cmd('vmware location check-trial-availability --location {loc} --sku {sku}')
-
-        # create a private cloud
-        self.cmd('vmware private-cloud create -g {rg} -n {privatecloud} --location {loc} --sku {sku} --cluster-size 4 --network-block 192.168.48.0/22 --nsxt-password 5rqdLj4GF3cePUe6( --vcenter-password UpfBXae9ZquZSDXk( --accept-eula')
-
-        self.cmd('vmware private-cloud list -g {rg}', checks=[
-            self.check('length(@)', 1)
-        ])
-
-        # update private cloud to changed default cluster size
-        private_cloud = self.cmd('vmware private-cloud update -g {rg} -n {privatecloud} --cluster-size 3').get_output_in_json()
-        self.kwargs["hosts"] = private_cloud["managementCluster"]["hosts"][:3]
-
-        # update private cloud to enable internet
-        self.cmd('vmware private-cloud update -g {rg} -n {privatecloud} --internet Enabled')
-
-        # create authorization
-        self.cmd('vmware authorization create -g {rg} -c {privatecloud} -n myauthname')
-
-        # delete authorization
-        self.cmd('vmware authorization delete -g {rg} -c {privatecloud} -n myauthname --yes')
+        # delete all extended network blocks
+        self.cmd('vmware private-cloud update -g {rg} -n {privatecloud} --ext-nw-blocks null')
 
         # set managed identity
         self.cmd('vmware private-cloud identity assign -g {rg} -c {privatecloud} --system-assigned')
-
-        # show managed identity
-        self.kwargs["object_id"] = self.cmd('vmware private-cloud identity show -g {rg} -c {privatecloud}').get_output_in_json()["principalId"]
-
-        self.cmd("keyvault create --name {kv} -g {rg} --sku standard --location {loc}")
-
-        self.cmd('keyvault key create --vault-name {kv} --name {key_vault_key} --protection software')
-
-        self.kwargs["vault_url"] = self.cmd('keyvault set-policy -g {rg} --name {kv} --key-permissions all --object-id {object_id}').get_output_in_json()['properties']['vaultUri']
 
         # enable cmk encryption
         self.cmd('vmware private-cloud enable-cmk-encryption -c {privatecloud} -g {rg} --enc-kv-key-name {key_vault_key} --enc-kv-url {vault_url}')
@@ -163,7 +111,12 @@ class VmwareScenarioTest(ScenarioTest):
 
         # delete the private cloud
         self.cmd('vmware private-cloud delete -g {rg} -n {privatecloud} --yes')
-        self.cmd('vmware private-cloud list -g {rg}', checks=[
-            self.check('length(@)', 0)
-        ])
-        self.cmd('group delete --name {rg} --yes --no-wait')
+
+        # cluster delete
+        self.cmd('vmware cluster delete -g {rg} -c {privatecloud} -n {cluster} --yes')
+
+        # cluster list zone
+        self.cmd('vmware cluster list-zones -g {rg} -c {privatecloud} -n {cluster}')
+
+        # delete the private cloud
+        self.cmd('vmware private-cloud delete -g {rg} -n {privatecloud} --yes')