+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,17 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+0.5.161
++++++++
+* Support `premium` cluster sku tier in `az aks create` and `az aks update` commands
+* Add option `--k8s-support-plan` to `az aks create` and `az aks update` commands
+* Add `az aks machine list` command to fetch list of machines in an agentpool.
+* Add `az aks machine show` command to fetch information about a specific machine in an agentpool.
+
+0.5.160
++++++++
+* Custom ips and managed ips can be assigned to aks cluster outbound resources
+
 0.5.159
 +++++++
 * Revert `az aks copilot` Command